chore: Simplify Accessibility integration by removing unnecessary abstractions
