+++ crates/renderer/src/accessibility.rs
@@ -3,48 +3,47 @@ use accesskit_winit::Adapter;
 use freya_common::EventMessage;
 use freya_core::{
     prelude::{
-        AccessibilityFocusDirection, AccessibilityProvider, AccessibilityState,
-        SharedAccessibilityState, ROOT_ID,
+        AccessibilityFocusDirection, AccessibilityManager, SharedAccessibilityManager, ROOT_ID,
     },
     types::FocusSender,
 };
 use winit::{event::WindowEvent, event_loop::EventLoopProxy, window::Window};
 
 /// Manages the accessibility integration with Accesskit.
-pub struct NativeAccessibility {
-    accessibility_state: SharedAccessibilityState,
+pub struct AccessKitManager {
+    accessibility_manager: SharedAccessibilityManager,
     accessibility_adapter: Adapter,
 }
 
-impl NativeAccessibility {
+impl AccessKitManager {
     pub fn new(window: &Window, proxy: EventLoopProxy<EventMessage>) -> Self {
         let title = window.title();
-        let accessibility_state = AccessibilityState::new().wrap();
+        let accessibility_manager = AccessibilityManager::new().wrap();
         let accessibility_adapter = {
-            let accessibility_state = accessibility_state.clone();
+            let accessibility_manager = accessibility_manager.clone();
             Adapter::new(
                 window,
                 move || {
-                    let mut accessibility_state = accessibility_state.lock().unwrap();
-                    accessibility_state.process(ROOT_ID, title.as_str())
+                    let mut accessibility_manager = accessibility_manager.lock().unwrap();
+                    accessibility_manager.process(ROOT_ID, title.as_str())
                 },
                 proxy,
             )
         };
         Self {
-            accessibility_state,
+            accessibility_manager,
             accessibility_adapter,
         }
     }
 
-    pub fn accessibility_state(&self) -> &SharedAccessibilityState {
-        &self.accessibility_state
+    pub fn accessibility_manager(&self) -> &SharedAccessibilityManager {
+        &self.accessibility_manager
     }
 
     /// Focus a new accessibility node
     pub fn set_accessibility_focus(&mut self, id: AccessibilityId) {
         let tree = self
-            .accessibility_state
+            .accessibility_manager
             .lock()
             .unwrap()
             .set_focus_with_update(Some(id));
@@ -60,13 +59,13 @@ impl NativeAccessibility {
 
     /// Remove the accessibility nodes
     pub fn clear_accessibility(&mut self) {
-        self.accessibility_state.lock().unwrap().clear();
+        self.accessibility_manager.lock().unwrap().clear();
     }
 
     /// Process the accessibility nodes
     pub fn render_accessibility(&mut self, title: &str) {
         let tree = self
-            .accessibility_state
+            .accessibility_manager
             .lock()
             .unwrap()
             .process(ROOT_ID, title);
@@ -80,7 +79,7 @@ impl NativeAccessibility {
         focus_sender: &FocusSender,
     ) {
         let tree = self
-            .accessibility_state
+            .accessibility_manager
             .lock()
             .unwrap()
             .set_focus_on_next_node(direction, focus_sender);
