+++ crates/renderer/src/app.rs
@@ -20,7 +20,7 @@ use uuid::Uuid;
 use winit::event::WindowEvent;
 use winit::{dpi::PhysicalSize, event_loop::EventLoopProxy};
 
-use crate::accessibility::NativeAccessibility;
+use crate::accessibility::AccessKitManager;
 use crate::{FontsConfig, HoveredNode, WindowEnv};
 
 fn winit_waker(proxy: &EventLoopProxy<EventMessage>) -> std::task::Waker {
@@ -61,7 +61,7 @@ pub struct App<State: 'static + Clone> {
     focus_sender: FocusSender,
     focus_receiver: FocusReceiver,
 
-    accessibility: NativeAccessibility,
+    accessibility: AccessKitManager,
 
     font_collection: FontCollection,
 
@@ -82,7 +82,7 @@ impl<State: 'static + Clone> App<State> {
         fonts_config: FontsConfig,
         mut plugins: PluginsManager,
     ) -> Self {
-        let accessibility = NativeAccessibility::new(&window_env.window, proxy.clone());
+        let accessibility = AccessKitManager::new(&window_env.window, proxy.clone());
 
         window_env.window_mut().set_visible(true);
 
@@ -283,7 +283,7 @@ impl<State: 'static + Clone> App<State> {
             layers,
             &layout,
             rdom,
-            &mut *self.accessibility.accessibility_state().lock().unwrap(),
+            &mut self.accessibility.accessibility_manager().lock().unwrap(),
         );
     }
 
@@ -348,7 +348,7 @@ impl<State: 'static + Clone> App<State> {
         &mut self.window_env
     }
 
-    pub fn accessibility(&mut self) -> &mut NativeAccessibility {
+    pub fn accessibility(&mut self) -> &mut AccessKitManager {
         &mut self.accessibility
     }
